requirements.txt
flask
mysql-connector-python
pyclean
